ref(tracing): Remove updated CLS from web-vitals
Switch to using the new definition of CLS and stop tracking both the old
and new definition.
